Added services folder, moved some files, and renamed a file
