+++ scanner.py
@@ -2,8 +2,8 @@
 import math
 import itertools
 from utils.contrast_utils import contrast_ratio, hex_to_rgb, css_to_hex
-from utils.css_parser import parse_css
-from utils.html_utils import parse_html, get_computed_style, has_direct_contents
+from services.css_parser import parse_css
+from services.html_parser import parse_html, get_computed_style, has_direct_contents
 from utils.debug import debug_print
 
 NORMAL_TEXT_CONTRAST_RAIO = 4.5
